+++ custom_components/duux/__init__.py
@@ -30,6 +30,7 @@ PLATFORMS = [
     Platform.SWITCH,
     Platform.SELECT,
     Platform.SENSOR,
+    Platform.BINARY_SENSOR,
 ]