+++ custom_components/duux/const.py
@@ -1,4 +1,5 @@
 # custom_components/duux/const.py
+from enum import Enum
 
 DOMAIN = "duux"
 CONF_EMAIL = "email"
@@ -70,3 +71,16 @@ DUUX_SUPPORTED_TYPES = (
 #         },
 #     ],
 # }
+
+# Error codes
+class DUUX_ERRID(Enum):
+    Unavailable = None
+    OK = 0
+    Ice_Detected = 4
+    Water_Tank_Full = 8
+    
+    Unknown_Error=9999999
+    
+    @classmethod
+    def _missing_(cls, value):
+        return cls.Unknown_Error